requirements.txt
flask
flask-cors
pydantic
scikit-learn
joblib
numpy
opencv-python-headless
mediapipe
groq
python-dotenv
fastapi
uvicorn[standard]
python-multipart
mediapipe
opencv-python-headless
numpy
scikit-learn
joblib
groq
python-dotenv
requests
